requirements.txt
boto3==1.38.32
botocore==1.38.32
click==8.1.8
dlib==20.0.0
face-recognition==1.3.0
face-recognition-models==0.3.0
jmespath==1.0.1
numpy==2.0.2
opencv-python==4.11.0.86
pillow==11.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
s3transfer==0.13.0
six==1.17.0
urllib3==1.26.20
